+++ net-reporter/src/query/charts/bandwidth_per_endpoint/request/requester.rs
@@ -1,5 +1,6 @@
 use std::sync::Arc;
 
+use net_reporter_api::api::network_bandwidth_per_endpoint::network_bandwidth_per_endpoint_filters::NetworkBandwidthPerEndpointFiltersDTO;
 use sqlx::types::chrono::DateTime;
 use sqlx::types::chrono::TimeZone;
 use sqlx::types::chrono::Utc;
@@ -19,27 +20,66 @@ use crate::query::charts::bandwidth_per_endpoint::response::network_bandwidth_pe
 use crate::query::charts::bandwidth_per_endpoint::response::endpoint::EndpointResponse;
 use crate::query::requester::Requester;
 
+
+const EXCLUDE_PROTOCOLS_FILTER_QUERY: &str = "
+    AND not (ARRAY(SELECT DISTINCT unnest(string_to_array(protocols, ':'))) && {})
+";
+
+const INCLUDE_PROTOCOLS_FILTER_QUERY: &str = "
+    AND ARRAY(SELECT DISTINCT unnest(string_to_array(protocols, ':'))) @> {}
+";
+
+const INCLUDE_ENDPOINT_FILTER_QUERY: &str = "
+    AND (COALESCE(lhs.id, rhs.id) IN (SELECT unnest({})))
+";
+
+const EXCLUDE_ENDPOINT_FILTER_QUERY: &str = "
+    AND (COALESCE(lhs.id, rhs.id) NOT IN (SELECT unnest({})))
+";
+
+const SET_LOWER_BYTES_BOUND: &str = "
+    AND LEAST(COALESCE(lhs.bytes_sent, 0), COALESCE(rhs.bytes_received, 0)) >= {}
+";
+
+const SET_UPPER_BYTES_BOUND: &str = "
+    AND GREATEST(COALESCE(lhs.bytes_sent, 0), COALESCE(rhs.bytes_received, 0)) < {}
+";
+
 const NETWORK_BANDWIDTH_PER_ENDPOINT_REQUEST_QUERY: &str = "
     SELECT
-    COALESCE(lhs.id, rhs.id) AS id,
-    lhs.bytes_sent AS bytes_sent,
-    rhs.bytes_received AS bytes_received
+        COALESCE(lhs.id, rhs.id) AS id,
+        COALESCE(lhs.bytes_sent, 0) AS bytes_sent,
+        COALESCE(rhs.bytes_received, 0) AS bytes_received
     FROM
     (
         SELECT
             src_addr AS id,
             SUM(packet_length) AS bytes_sent
         FROM bandwidth_per_endpoint_aggregate
-        WHERE group_id = $1 AND bucket >= $2 AND bucket < $3
+        WHERE 
+            group_id = $1
+            AND bucket >= $2
+            AND bucket < $3
+            {}
         GROUP BY src_addr
     ) AS lhs full outer join (
         SELECT
             dst_addr AS id,
             SUM(packet_length) AS bytes_received
         FROM bandwidth_per_endpoint_aggregate
-        WHERE group_id = $1 AND bucket >= $2 AND bucket < $3
+        WHERE 
+            group_id = $1
+            AND bucket >= $2 
+            AND bucket < $3
+            {}
         GROUP BY dst_addr
-    ) AS rhs ON lhs.id = rhs.id;
+    ) AS rhs ON lhs.id = rhs.id
+    WHERE
+        1 = 1
+        {}
+        {}
+        {}
+    ORDER BY id;
 ";
 
 #[derive(Default)]
@@ -50,17 +90,99 @@ impl NetworkBandwidthPerEndpointRequester {
         Box::new(self)
     }
 
+    async fn get_query_based_on_requested_filters(filters: &NetworkBandwidthPerEndpointFiltersDTO) -> String {
+        let mut placeholder_value = 4;
+        let mut request_query = NETWORK_BANDWIDTH_PER_ENDPOINT_REQUEST_QUERY.to_owned();
+
+        match filters.is_include_protocols_mode() {
+            Some(true) => {
+                let protocols_query = INCLUDE_PROTOCOLS_FILTER_QUERY.to_string().replace("{}", format!("${}", placeholder_value).as_str());
+                placeholder_value += 1;
+                request_query = request_query.replacen("{}", protocols_query.as_str(), 2);
+            },
+            Some(false) => {
+                let protocols_query = EXCLUDE_PROTOCOLS_FILTER_QUERY.to_string().replace("{}", format!("${}", placeholder_value).as_str());
+                placeholder_value += 1;
+                request_query = request_query.replacen("{}", protocols_query.as_str(), 2);
+            },
+            None => request_query = request_query.replacen("{}", "", 2)
+        }
+
+        match filters.is_include_endpoints_mode() {
+            Some(true) => {
+                let endpoints_query = INCLUDE_ENDPOINT_FILTER_QUERY.to_string().replace("{}", format!("${}", placeholder_value).as_str());
+                placeholder_value += 1;
+                request_query = request_query.replacen("{}", endpoints_query.as_str(), 1);
+            },
+            Some(false) => { 
+                let endpoints_query = EXCLUDE_ENDPOINT_FILTER_QUERY.to_string().replace("{}", format!("${}", placeholder_value).as_str());
+                placeholder_value += 1;
+                request_query = request_query.replacen("{}", endpoints_query.as_str(), 1) 
+            },
+            None => request_query = request_query.replacen("{}", "", 1)
+        };
+
+        match filters.get_bytes_lower_bound() {
+            Some(_) => {
+                let lower_bytes_query = SET_LOWER_BYTES_BOUND.to_string().replace("{}", format!("${}", placeholder_value).as_str());
+                placeholder_value += 1;
+                request_query = request_query.replacen("{}", lower_bytes_query.as_str(), 1)
+            },
+            None => request_query = request_query.replacen("{}", "", 1)
+        };
+
+        match filters.get_bytes_upper_bound() {
+            Some(_) => {
+                let upper_bytes_query = SET_UPPER_BYTES_BOUND.to_string().replace("{}", format!("${}", placeholder_value).as_str());
+                request_query = request_query.replacen("{}", upper_bytes_query.as_str(), 1)
+            },
+            None => request_query = request_query.replacen("{}", "", 1)
+        };
+        request_query
+    }
+
     async fn execute_query(
         connection_pool: Arc<Pool<Postgres>>,
         group_id: Option<&str>,
         start_date: DateTime<Utc>,
         end_date: DateTime<Utc>,
+        filters: &NetworkBandwidthPerEndpointFiltersDTO,
     ) -> Result<Vec<EndpointResponse>, Error> {
-        sqlx::query_as(NETWORK_BANDWIDTH_PER_ENDPOINT_REQUEST_QUERY)
+        let request_query = Self::get_query_based_on_requested_filters(filters).await;
+
+        log::debug!("Request query: {}", request_query);
+        log::debug!("Group id: {:?}", group_id);
+        log::debug!("Start date: {}", start_date);
+        log::debug!("End date: {}", end_date);
+        log::debug!("Filters: {:?}", filters);
+
+
+        let mut sqlx_query = sqlx::query_as(request_query.as_str())
             .bind(group_id)
             .bind(start_date)
-            .bind(end_date)
-            .fetch_all(connection_pool.as_ref())
+            .bind(end_date);
+
+        sqlx_query = match filters.is_include_protocols_mode() {
+            Some(_) => sqlx_query.bind(filters.get_protocols()),
+            None => sqlx_query,
+        };
+
+        sqlx_query = match filters.is_include_endpoints_mode() {
+            Some(_) => sqlx_query.bind(filters.get_endpoints()),
+            None => sqlx_query,
+        };
+
+        sqlx_query = match filters.get_bytes_lower_bound() {
+            Some(lower_bound) => sqlx_query.bind(lower_bound),
+            None => sqlx_query,
+        };
+
+        sqlx_query = match filters.get_bytes_upper_bound() {
+            Some(upper_bound) => sqlx_query.bind(upper_bound),
+            None => sqlx_query,
+        };
+
+        sqlx_query.fetch_all(connection_pool.as_ref())
             .await
     }
 }
@@ -81,12 +203,14 @@ impl Requester for NetworkBandwidthPerEndpointRequester {
         let request = NetworkBandwidthPerEndpointRequestDTO::decode(enveloped_request.get_data());
         let request_start_date: DateTime<Utc> = Utc.timestamp_millis_opt(request.get_start_date_time()).unwrap();
         let request_end_date: DateTime<Utc> = Utc.timestamp_millis_opt(request.get_end_date_time()).unwrap();
+        let filters = request.get_filters();
 
         let executed_query_response = Self::execute_query(
             connection_pool,
             request_group_id,
             request_start_date,
-            request_end_date
+            request_end_date,
+            filters,
         ).await;
 
         if let Err(e) = executed_query_response {